--- runtime-light/crypto/crypto.cmake
@@ -1,3 +1,0 @@
-prepend(RUNTIME_CRYPTO_SRC crypto/
-        crypto-builtins.cpp
-)
