+++ runtime-light/runtime-light.cmake
@@ -1,6 +1,5 @@
 include(${RUNTIME_LIGHT_DIR}/allocator/allocator.cmake)
 include(${RUNTIME_LIGHT_DIR}/core/core.cmake)
-include(${RUNTIME_LIGHT_DIR}/crypto/crypto.cmake)
 include(${RUNTIME_LIGHT_DIR}/scheduler/scheduler.cmake)
 include(${RUNTIME_LIGHT_DIR}/stdlib/stdlib.cmake)
 include(${RUNTIME_LIGHT_DIR}/streams/streams.cmake)
@@ -11,7 +10,6 @@ include(${RUNTIME_LIGHT_DIR}/memory-resource-impl/memory-resource-impl.cmake)
 
 set(RUNTIME_LIGHT_SRC
     ${RUNTIME_CORE_SRC}
-    ${RUNTIME_CRYPTO_SRC}
     ${RUNTIME_STDLIB_SRC}
     ${RUNTIME_SCHEDULER_SRC}
     ${RUNTIME_ALLOCATOR_SRC}
